requirements.txt
# ─── Core ML ─────────────────────────────────────────
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0

# ─── Image Processing ────────────────────────────────
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ─── Augmentation ────────────────────────────────────
albumentations>=1.3.1

# ─── Visualization ───────────────────────────────────
supervision>=0.17.0
matplotlib>=3.7.0

# ─── Dataset Management ─────────────────────────────
roboflow>=0.2.0
pyyaml>=6.0

# ─── Testing ─────────────────────────────────────────
pytest>=7.4.0

# ─── Optional: Colab / Notebook ─────────────────────
# ipywidgets>=8.0.0
# jupyter>=1.0.0
